+++ .eslintrc.js
@@ -1,23 +1,25 @@
 const typescriptEslintRecommended = require('@typescript-eslint/eslint-plugin').configs.recommended;
 
 module.exports = {
-  parser: 'babel-eslint',
+  parser: '@typescript-eslint/parser',
   parserOptions: {
-    ecmaVersion: 2018,
+    ecmaVersion: 2022,
     sourceType: 'module',
+    warnOnUnsupportedTypeScriptVersion: false,
     ecmaFeatures: {
       modules: true,
       experimentalObjectRestSpread: true,
     },
   },
   extends: [
+    'eslint:recommended',
     'plugin:@typescript-eslint/recommended',
     // Enables eslint-plugin-prettier and eslint-config-prettier.
     // This will display prettier errors as ESLint errors.
     // Make sure this is always the last configuration in the extends array.
     'plugin:prettier/recommended',
   ],
-  plugins: ['import', 'babel', 'prettier'],
+  plugins: ['import', 'prettier', '@typescript-eslint'],
   rules: {
     'prettier/prettier': 'warn',
     'no-console': 'warn',
@@ -31,6 +33,7 @@ module.exports = {
       parser: '@typescript-eslint/parser',
       parserOptions: {
         project: './tsconfig.json',
+        warnOnUnsupportedTypeScriptVersion: false,
       },
       plugins: ['@typescript-eslint'],
       rules: Object.assign(typescriptEslintRecommended.rules, {
@@ -38,78 +41,6 @@ module.exports = {
         '@typescript-eslint/explicit-function-return-type': 'off',
         '@typescript-eslint/no-explicit-any': 'off',
         '@typescript-eslint/no-var-requires': 'off',
-        '@typescript-eslint/no-empty-function': 'off',
-        'no-console': 'off',
-        'import/order': [
-          'error',
-          {
-            groups: ['type', 'object', 'builtin', 'external', 'internal', 'parent', 'sibling', 'index'],
-            'newlines-between': 'always',
-            pathGroups: [
-              {
-                pattern: '~',
-                group: 'unknown',
-                position: 'before',
-              },
-              {
-                pattern: '@mysql/xdevapi',
-                group: 'external',
-                position: 'before',
-              },
-              {
-                pattern: 'knex',
-                group: 'external',
-                position: 'before',
-              },
-              {
-                pattern: 'pg',
-                group: 'external',
-                position: 'before',
-              },
-              {
-                pattern: 'oracledb',
-                group: 'external',
-                position: 'before',
-              },
-              {
-                pattern: 'lib/client/**',
-                group: 'internal',
-                position: 'before',
-              },
-              {
-                pattern: 'lib/controllers/**',
-                group: 'internal',
-                position: 'before',
-              },
-              {
-                pattern: 'lib/errors/**',
-                group: 'internal',
-                position: 'before',
-              },
-              {
-                pattern: 'lib/middleware/**',
-                group: 'internal',
-                position: 'before',
-              },
-              {
-                pattern: 'lib/types/**',
-                group: 'internal',
-                position: 'before',
-              },
-              {
-                pattern: 'lib/utils/**',
-                group: 'internal',
-                position: 'before',
-              },
-            ],
-            pathGroupsExcludedImportTypes: ['mysql', '@mysql/xdevapi', 'knex', 'pg', 'oracledb', 'unknown'],
-            alphabetize: {
-              order: 'asc',
-              caseInsensitive: true,
-            },
-            warnOnUnassignedImports: true,
-          },
-        ],
       }),
     },
   ],