+++ examples/mysqlx/src/schemas/employee.ts
@@ -1,4 +1,4 @@
-import { IDocumentSchema, IDocumentValidation, IDocument, getDocument, generateId } from 'crud-node';
+import { IDocument, IDocumentSchema, IDocumentValidation, generateId, getDocument } from 'crud-node';
 
 export enum EmployeeProps {
   _id = '_id',